requirements.txt
aiosqlite==0.22.1
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.6
asgi-lifespan==2.1.0
asttokens==3.0.1
asyncpg==0.31.0
attrs==25.4.0
beartype==0.22.9
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
coolname==2.2.0
cryptography==46.0.3
dateparser==1.2.2
debugpy==1.8.19
decorator==5.2.1
docker==7.1.0
dotenv==0.9.9
exceptiongroup==1.3.1
executing==2.2.1
fakeredis==2.33.0
fastapi==0.128.0
fsspec==2026.1.0
graphviz==0.21
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
lupa==2.6
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.4.1
oauthlib==3.3.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==1.0.3
pendulum==3.1.0
pexpect==4.9.0
platformdirs==4.5.1
pluggy==1.6.0
prefect==3.6.11
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.6
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
readchar==4.2.1
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sodapy==2.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
text-unidecode==1.3
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
